+++ src/rust/src/rlib.rs
@@ -250,6 +250,12 @@ pub fn clone_robj(robj: Robj) -> Robj {
 fn test_robj_to_usize(robj: Robj) -> RResult<String> {
     robj_to!(usize, robj).map(rdbg)
 }
+
+#[extendr]
+fn test_robj_to_f64(robj: Robj) -> RResult<String> {
+    robj_to!(f64, robj).map(rdbg)
+}
+
 #[extendr]
 fn test_robj_to_i64(robj: Robj) -> RResult<String> {
     robj_to!(i64, robj).map(rdbg)
@@ -258,6 +264,11 @@ fn test_robj_to_i64(robj: Robj) -> RResult<String> {
 #[extendr]
 fn test_robj_to_u32(robj: Robj) -> RResult<String> {
     robj_to!(u32, robj).map(rdbg)
+}
+
+#[extendr]
+fn test_robj_to_i32(robj: Robj) -> RResult<String> {
+    robj_to!(i32, robj).map(rdbg)
 }
 
 #[extendr]
@@ -303,8 +314,10 @@ extendr_module! {
     fn clone_robj;
 
     fn test_robj_to_usize;
+    fn test_robj_to_f64;
     fn test_robj_to_i64;
     fn test_robj_to_u32;
+    fn test_robj_to_i32;
     fn test_print_string;
     fn test_robj_to_expr;
     fn test_wrong_call_pl_lit;